requirements.txt
# Core dependencies
pymongo==4.6.0
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# NLP and ML dependencies
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Scheduling and utilities
apscheduler==3.10.4
python-dotenv==1.0.0
hashlib2==1.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0